Fix line number in broken test.
We have this code:
20: ddd :: Num a => a -> a -> a
21: ddd vv ww = vv +! ww

The intention was to ask for the type definition of symbol "a" in line 20, and
then assert that no type definitions were found.

The reality is that the test was asking for the definition of the symbol at (20,
15) in 0-based indexing, which is the "!" in "+!". Until recently, ghcide could
not find type definitions for "+!", so no type definitions were
found and the test passed. But now, ghcide can find type definitions for "+!",
and this test has begun to fail.

The solution is to change (20, 15) to (19, 15), so that we ask for the type
definitions of the symbol "a", which will not be found.